clip_api/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
black==24.4.2
blinker==1.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
coloredlogs==15.0.1
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
flake8==7.0.0
flatbuffers==24.3.25
fsspec==2024.5.0
ftfy==6.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
importlib-metadata==4.6.4
isort==5.13.2
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.3
nudenet==3.0.8
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
onnxruntime==1.18.0
open-clip-torch==2.24.0
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
orjson==3.10.3
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
protobuf==5.27.0
pycodestyle==2.11.1
pydantic==2.7.2
pydantic_core==2.18.3
pyflakes==3.2.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==5.4.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
SecretStorage==3.3.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12.1
systemd-python==234
timm==1.0.3
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
triton==2.3.0
typer==0.12.3
typing_extensions==4.12.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
zipp==1.0.0
